wip: add buttons - refactor and unify scan and style
