+++ src/ui/components/Scanner/Scanner.scss
@@ -1,27 +1,10 @@
-.qr-code-scanner-full-page {
-  ion-header ion-toolbar {
-    margin-top: 0.625rem;
-
-    .back-button {
-      width: 2.5rem;
-      height: 2.5rem;
-      border-radius: 1.5rem;
-      background: var(--ion-color-light-grey);
-      margin-left: 1.25rem;
-    }
-
-    ion-icon {
-      font-size: 1.85em !important;
-    }
-  }
-}
-
-.qr-code-scanner {
+ion-grid.qr-code-scanner {
   --scale-scan: 1;
   height: calc(var(--scale-h) * 100);
   display: flex;
   flex-direction: column;
-  justify-content: center;
+  justify-content: space-around;
+  padding-inline: 1.25rem;
 
   ion-row {
     display: flex;
@@ -44,7 +27,13 @@
     }
   }
 
+  .page-footer .secondary-button {
+    opacity: 0.5;
+    backdrop-filter: blur(0.33rem);
+  }
+
   @media screen and (min-width: 250px) and (max-width: 370px) {
     --scale-scan: 0.7;
+    padding-inline: 0.9rem;
   }
 }
