+++ src/ui/pages/FullPageScanner/FullPageScanner.scss
@@ -1,10 +1,20 @@
 .qr-code-scanner-full-page {
-  &.responsive-page-layout {
-    background: transparent;
+  background: transparent;
 
-    ion-toolbar {
-      margin-top: 0;
-      --background: transparent;
+  ion-header ion-toolbar {
+    margin-top: 0.625rem;
+    --background: transparent;
+
+    .back-button {
+      width: 2.5rem;
+      height: 2.5rem;
+      border-radius: 1.5rem;
+      background: var(--ion-color-light-grey);
+      margin-left: 1.25rem;
+    }
+
+    ion-icon {
+      font-size: 1.85em !important;
     }
   }
 
@@ -18,7 +28,6 @@
     }
 
     .qr-code-scanner-icon {
-      padding-bottom: 4rem;
       margin: 0;
     }
