+++ src/ui/pages/FullPageScanner/FullPageScanner.tsx
@@ -1,19 +1,8 @@
 import { arrowBackOutline } from "ionicons/icons";
-import {
-  IonPage,
-  IonToolbar,
-  IonButtons,
-  IonButton,
-  IonIcon,
-  IonHeader,
-} from "@ionic/react";
 import { useEffect, useRef } from "react";
 import { Scanner } from "../../components/Scanner";
-import {
-  getCurrentOperation,
-  setCurrentOperation,
-} from "../../../store/reducers/stateCache";
-import { useAppDispatch, useAppSelector } from "../../../store/hooks";
+import { setCurrentOperation } from "../../../store/reducers/stateCache";
+import { useAppDispatch } from "../../../store/hooks";
 import {
   FullPageScannerProps,
   ScannerRefComponent,
@@ -22,13 +11,10 @@ import { OperationType } from "../../globals/types";
 import "./FullPageScanner.scss";
 import { ResponsivePageLayout } from "../../components/layout/ResponsivePageLayout";
 import { PageHeader } from "../../components/PageHeader";
-import { PageFooter } from "../../components/PageFooter";
-import { i18n } from "../../../i18n";
 
 const FullPageScanner = ({ showScan, setShowScan }: FullPageScannerProps) => {
   const pageId = "qr-code-scanner-full-page";
   const dispatch = useAppDispatch();
-  const currentOperation = useAppSelector(getCurrentOperation);
   const scannerRef = useRef<ScannerRefComponent>(null);
 
   useEffect(() => {
@@ -43,14 +29,6 @@ const FullPageScanner = ({ showScan, setShowScan }: FullPageScannerProps) => {
     document
       ?.querySelector("body.scanner-active > div:last-child")
       ?.classList.add("hide");
-  };
-
-  const handlePrimaryButtonAction = () => {
-    //
-  };
-
-  const handleSecondaryButtonAction = () => {
-    //
   };
 
   return (
@@ -69,24 +47,6 @@ const FullPageScanner = ({ showScan, setShowScan }: FullPageScannerProps) => {
         ref={scannerRef}
         handleReset={handleReset}
       />
-      {currentOperation === OperationType.MULTI_SIG_INITIATOR_SCAN && (
-        <PageFooter
-          pageId={pageId}
-          primaryButtonText={`${i18n.t("createidentifier.scan.initiate")}`}
-          primaryButtonAction={() => handlePrimaryButtonAction()}
-          secondaryButtonText={`${i18n.t("createidentifier.scan.pasteoobi")}`}
-          secondaryButtonAction={() => handleSecondaryButtonAction()}
-        />
-      )}
-      {currentOperation === OperationType.MULTI_SIG_RECEIVER_SCAN && (
-        <PageFooter
-          pageId={pageId}
-          secondaryButtonText={`${i18n.t(
-            "createidentifier.scan.pastecontents"
-          )}`}
-          secondaryButtonAction={() => handleSecondaryButtonAction()}
-        />
-      )}
     </ResponsivePageLayout>
   );
 };
