+++ src/ui/pages/Scan/Scan.scss
@@ -1,8 +1,12 @@
 .scan-tab {
   padding: 0;
 
-  .qr-code-scanner-icon {
-    padding-bottom: 6.7rem;
-    margin: 0;
+  ion-grid {
+    padding-top: 10vh;
+    padding-bottom: 10vh;
+
+    .qr-code-scanner-icon {
+      margin: 0;
+    }
   }
 }
